+++ lib/stream.dart
@@ -7,11 +7,14 @@ library stream;
 import "dart:io";
 import "dart:async";
 import "dart:collection";
+import "dart:mirrors";
+import 'package:meta/meta.dart';
 import 'package:args/args.dart' show Options;
 import "package:logging/logging.dart" show Logger;
 
 import "package:rikulo_commons/util.dart";
 import "package:rikulo_commons/io.dart";
+import "package:rikulo_commons/mirrors.dart";
 
 import "plugin.dart";