+++ CHANGES.md
@@ -1,5 +1,9 @@
 #CHANGES
 
+**0.5.2**
+
+* The error mapping takes the syntax of Map<code or exception, uri or function>.
+
 **0.5.1**
 
 * The comment tag is renamed to a pair of `[!--` and `--]`